+++ flowman-spark-extensions/src/test/scala/com/dimajix/spark/sql/catalyst/SqlBuilderTest.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 The Flowman Authors
+ * Copyright (C) 2018-2024 The Flowman Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,9 +24,6 @@ import com.dimajix.spark.testing.LocalSparkSession
 
 
 class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
-    implicit class NoQuotes(s:String) {
-        def withoutQuotes : String = s.replace("`", "")
-    }
     override def beforeAll(): Unit = {
         super.beforeAll()
 
@@ -66,35 +63,57 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         super.afterAll()
     }
 
-    behavior of "The SQLBuilder"
+    behavior of "The SqlBuilder"
 
     it should "support basic selects" in (if (hiveSupported) {
         val df1 = spark.sql("SELECT * FROM sql_builder_0")
         val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql1))
-        sql1.withoutQuotes should be ("SELECT col_0, col_1, ts FROM default.sql_builder_0")
+        sql1 should be ("SELECT `col_0`, `col_1`, `ts` FROM `default`.`sql_builder_0`")
 
         val df2 = spark.sql("SELECT CONCAT(col_0, col_1) AS result FROM sql_builder_0")
         val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql2))
-        sql2.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0")
+        sql2 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0`")
 
         val df3 = spark.sql("SELECT CONCAT(x.col_0, x.col_1) AS result FROM sql_builder_0 AS x")
         val sql3 = new SqlBuilder(df3.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql3))
-        sql3.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0")
+        sql3 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0`")
 
         val df4 = spark.sql("SELECT CONCAT(x.col_0, x.col_1) AS result FROM sql_builder_0 AS x WHERE x.col_0 = 67")
         val sql4 = new SqlBuilder(df4.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql4))
-        sql4.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 WHERE (col_0 = 67)")
+        sql4 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` WHERE (`col_0` = 67)")
+    })
+
+    it should "support basic arithmetics" in (if (hiveSupported) {
+        val df1 = spark.sql("SELECT 1+2")
+        val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql1))
+        sql1 should be("SELECT (1 + 2) AS `(1 + 2)`")
+
+        val df2 = spark.sql("SELECT 1+2 AS result")
+        val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql2))
+        sql2 should be("SELECT (1 + 2) AS `result`")
+
+        val df3 = spark.sql("SELECT -3*(1+2)")
+        val sql3 = new SqlBuilder(df3.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql3))
+        sql3 should be("SELECT (-3 * (1 + 2)) AS `(-3 * (1 + 2))`")
+
+        val df4 = spark.sql("SELECT -(8*3)-(1+2)")
+        val sql4 = new SqlBuilder(df4.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql4))
+        sql4 should be("SELECT ((- (8 * 3)) - (1 + 2)) AS `((- (8 * 3)) - (1 + 2))`")
     })
 
     it should "support VIEWs" in (if (hiveSupported) {
         val df1 = spark.sql("SELECT * FROM sql_builder_1")
         val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql1))
-        sql1.withoutQuotes should be ("SELECT col_0, col_1, ts FROM default.sql_builder_1")
+        sql1 should be ("SELECT `col_0`, `col_1`, `ts` FROM `default`.`sql_builder_1`")
     })
 
     it should "support JOINs" in (if (hiveSupported) {
@@ -108,7 +127,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.col_1 = '67'""".stripMargin)
         val sql5 = new SqlBuilder(df5.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql5))
-        sql5.withoutQuotes should be ("SELECT concat(CAST(gen_attr_1 AS STRING), gen_attr_2) AS result FROM (SELECT col_0 AS gen_attr_1, col_1 AS gen_attr_3, ts AS gen_attr_5 FROM default.sql_builder_0) AS gen_subquery_0 INNER JOIN (SELECT col_0 AS gen_attr_4, col_1 AS gen_attr_2, ts AS gen_attr_6 FROM default.sql_builder_0) AS gen_subquery_1 ON (gen_attr_1 = gen_attr_4) WHERE (gen_attr_3 = '67')")
+        sql5 should be ("SELECT concat(CAST(`gen_attr_1` AS STRING), `gen_attr_2`) AS `result` FROM (SELECT `col_0` AS `gen_attr_1`, `col_1` AS `gen_attr_3`, `ts` AS `gen_attr_5` FROM `default`.`sql_builder_0`) AS `gen_subquery_0` INNER JOIN (SELECT `col_0` AS `gen_attr_4`, `col_1` AS `gen_attr_2`, `ts` AS `gen_attr_6` FROM `default`.`sql_builder_0`) AS `gen_subquery_1` ON (`gen_attr_1` = `gen_attr_4`) WHERE (`gen_attr_3` = '67')")
     })
 
     it should "support UNIONs without column names" in (if (hiveSupported) {
@@ -127,9 +146,9 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         val sql3 = new SqlBuilder(df3.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql3))
         if (spark.version < "3.2")
-            sql3.withoutQuotes should be("SELECT concat(CAST(col_0 AS STRING), col_1) AS concat(CAST(col_0 AS STRING), col_1) FROM default.sql_builder_0 UNION ALL SELECT concat(CAST(col_0 AS STRING), col_1) AS concat(CAST(col_0 AS STRING), col_1) FROM default.sql_builder_0 WHERE (col_1 = '67')")
+            sql3 should be("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `concat(CAST(col_0 AS STRING), col_1)` FROM `default`.`sql_builder_0` UNION ALL SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `concat(CAST(col_0 AS STRING), col_1)` FROM `default`.`sql_builder_0` WHERE (`col_1` = '67')")
         else
-            sql3.withoutQuotes should be("SELECT concat(CAST(col_0 AS STRING), col_1) AS concat(col_0, col_1) FROM default.sql_builder_0 UNION ALL SELECT concat(CAST(col_0 AS STRING), col_1) AS concat(col_0, col_1) FROM default.sql_builder_0 WHERE (col_1 = '67')")
+            sql3 should be("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `concat(col_0, col_1)` FROM `default`.`sql_builder_0` UNION ALL SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `concat(col_0, col_1)` FROM `default`.`sql_builder_0` WHERE (`col_1` = '67')")
     })
 
     it should "support UNIONs with one attribute name" in (if (hiveSupported) {
@@ -147,7 +166,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.col_1 = '67'""".stripMargin)
         val sql4 = new SqlBuilder(df4.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql4))
-        sql4.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 UNION ALL SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 WHERE (col_1 = '67')")
+        sql4 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` UNION ALL SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` WHERE (`col_1` = '67')")
     })
 
     it should "support UNIONs with different column names" in (if (hiveSupported) {
@@ -165,7 +184,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.col_1 = '67'""".stripMargin)
         val sql5 = new SqlBuilder(df5.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql5))
-        sql5.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 UNION ALL SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 WHERE (col_1 = '67')")
+        sql5 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` UNION ALL SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` WHERE (`col_1` = '67')")
     })
 
     it should "support UNIONs with matching column names" in (if (hiveSupported) {
@@ -183,7 +202,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.col_1 = '67'""".stripMargin)
         val sql6 = new SqlBuilder(df6.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql6))
-        sql6.withoutQuotes should be ("SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 UNION ALL SELECT concat(CAST(col_0 AS STRING), col_1) AS result FROM default.sql_builder_0 WHERE (col_1 = '67')")
+        sql6 should be ("SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` UNION ALL SELECT concat(CAST(`col_0` AS STRING), `col_1`) AS `result` FROM `default`.`sql_builder_0` WHERE (`col_1` = '67')")
     })
 
     it should "support subquery UNIONs" in (if (hiveSupported) {
@@ -208,7 +227,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         )
         val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql1))
-        sql1.withoutQuotes should be ("SELECT gen_attr_1 AS first, gen_attr_3 AS second FROM (SELECT col_0 AS gen_attr_1, col_1 AS gen_attr_3 FROM default.sql_builder_0 UNION ALL SELECT col_0 AS gen_attr_1, col_1 AS gen_attr_3 FROM default.sql_builder_0) AS gen_subquery_2 WHERE (NOT (gen_attr_1 = 7))")
+        sql1 should be ("SELECT `gen_attr_1` AS `first`, `gen_attr_3` AS `second` FROM (SELECT `col_0` AS `gen_attr_1`, `col_1` AS `gen_attr_3` FROM `default`.`sql_builder_0` UNION ALL SELECT `col_0` AS `gen_attr_1`, `col_1` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_2` WHERE (NOT (`gen_attr_1` = 7))")
     })
 
     it should "support subquery UNIONs with partial column names" in (if (hiveSupported) {
@@ -232,7 +251,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         )
         val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql2))
-        sql2.withoutQuotes should be ("SELECT CAST(col_0 AS STRING) AS first, col_1 AS second FROM default.sql_builder_0 UNION ALL SELECT upper(CAST(col_0 AS STRING)) AS first, col_1 AS second FROM default.sql_builder_0")
+        sql2 should be ("SELECT CAST(`col_0` AS STRING) AS `first`, `col_1` AS `second` FROM `default`.`sql_builder_0` UNION ALL SELECT upper(CAST(`col_0` AS STRING)) AS `first`, `col_1` AS `second` FROM `default`.`sql_builder_0`")
     })
 
     it should "support subquery UNIONs with partial column names and filters" in (if (hiveSupported) {
@@ -257,7 +276,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         )
         val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql2))
-        sql2.withoutQuotes should be ("SELECT gen_attr_1 AS first, gen_attr_3 AS second FROM (SELECT CAST(col_0 AS STRING) AS gen_attr_1, col_1 AS gen_attr_3 FROM default.sql_builder_0 UNION ALL SELECT upper(CAST(col_0 AS STRING)) AS gen_attr_1, col_1 AS gen_attr_3 FROM default.sql_builder_0) AS gen_subquery_2 WHERE (NOT (CAST(gen_attr_1 AS INT) = 7))")
+        sql2 should be ("SELECT `gen_attr_1` AS `first`, `gen_attr_3` AS `second` FROM (SELECT CAST(`col_0` AS STRING) AS `gen_attr_1`, `col_1` AS `gen_attr_3` FROM `default`.`sql_builder_0` UNION ALL SELECT upper(CAST(`col_0` AS STRING)) AS `gen_attr_1`, `col_1` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_2` WHERE (NOT (CAST(`gen_attr_1` AS INT) = 7))")
     })
 
     it should "support subquery UNIONs without alias" in (if (hiveSupported) {
@@ -281,7 +300,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
         )
         val sql3 = new SqlBuilder(df3.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql3))
-        sql3.withoutQuotes should be ("SELECT CAST(col_0 AS STRING) AS first, col_1 AS second FROM default.sql_builder_0 UNION ALL SELECT upper(CAST(col_0 AS STRING)) AS first, col_1 AS second FROM default.sql_builder_0")
+        sql3 should be ("SELECT CAST(`col_0` AS STRING) AS `first`, `col_1` AS `second` FROM `default`.`sql_builder_0` UNION ALL SELECT upper(CAST(`col_0` AS STRING)) AS `first`, `col_1` AS `second` FROM `default`.`sql_builder_0`")
     })
 
     it should "support subquery WINDOW functions" in (if (hiveSupported) {
@@ -300,7 +319,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.rank = 1""".stripMargin)
         val sql7 = new SqlBuilder(df7.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql7))
-        sql7.withoutQuotes should be ("SELECT gen_attr_0 AS col_0, gen_attr_1 AS col_1 FROM (SELECT col_0 AS gen_attr_0, col_1 AS gen_attr_1, row_number() OVER (PARTITION BY col_0 ORDER BY ts ASC) AS gen_attr_2 FROM default.sql_builder_0) AS x WHERE (gen_attr_2 = 1)")
+        sql7 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1`, row_number() OVER (PARTITION BY `col_0` ORDER BY `ts` ASC) AS `gen_attr_2` FROM `default`.`sql_builder_0`) AS `x` WHERE (`gen_attr_2` = 1)")
     })
 
     it should "support plain WINDOW functions" in (if (hiveSupported) {
@@ -314,7 +333,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |""".stripMargin)
         val sql8 = new SqlBuilder(df8.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql8))
-        sql8.withoutQuotes should be ("SELECT col_0, col_1, lead(col_1, 1, NULL) OVER (PARTITION BY col_0 ORDER BY ts ASC) AS rank FROM default.sql_builder_0")
+        sql8 should be ("SELECT `col_0`, `col_1`, lead(`col_1`, 1, NULL) OVER (PARTITION BY `col_0` ORDER BY `ts` ASC) AS `rank` FROM `default`.`sql_builder_0`")
     })
 
     it should "support subquery WINDOW functions with range" in (if (hiveSupported) {
@@ -333,7 +352,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.rank = 1""".stripMargin)
         val sql9 = new SqlBuilder(df9.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql9))
-        sql9.withoutQuotes should be ("SELECT gen_attr_0 AS col_0, gen_attr_1 AS col_1 FROM (SELECT col_0 AS gen_attr_0, col_1 AS gen_attr_1, max(col_0) OVER (PARTITION BY col_1 ORDER BY ts ASC ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS gen_attr_2 FROM default.sql_builder_0) AS x WHERE (gen_attr_2 = 1)")
+        sql9 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1`, max(`col_0`) OVER (PARTITION BY `col_1` ORDER BY `ts` ASC ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS `gen_attr_2` FROM `default`.`sql_builder_0`) AS `x` WHERE (`gen_attr_2` = 1)")
     })
 
     it should "support subquery WINDOW functions with DESC NULLS FIRST" in (if (hiveSupported) {
@@ -352,7 +371,42 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |WHERE x.rank = 1""".stripMargin)
         val sql10 = new SqlBuilder(df10.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql10))
-        sql10.withoutQuotes should be ("SELECT gen_attr_0 AS col_0, gen_attr_1 AS col_1 FROM (SELECT col_0 AS gen_attr_0, col_1 AS gen_attr_1, max(col_0) OVER (PARTITION BY col_1 ORDER BY ts DESC NULLS FIRST ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS gen_attr_2 FROM default.sql_builder_0) AS x WHERE (gen_attr_2 = 1)")
+        sql10 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1`, max(`col_0`) OVER (PARTITION BY `col_1` ORDER BY `ts` DESC NULLS FIRST ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS `gen_attr_2` FROM `default`.`sql_builder_0`) AS `x` WHERE (`gen_attr_2` = 1)")
+    })
+
+    it should "support GROUP BY expressions" in (if (hiveSupported) {
+        val df8 = spark.sql(
+            """
+              |SELECT
+              |   col_1,
+              |   sum(col_0) AS sum
+              |FROM sql_builder_0
+              |GROUP BY col_1
+              |""".stripMargin)
+        val sql8 = new SqlBuilder(df8.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql8))
+        if (spark.version < "3.2")
+            sql8 should be ("SELECT `gen_attr_0` AS `col_1`, sum(CAST(`gen_attr_2` AS BIGINT)) AS `sum` FROM (SELECT `col_0` AS `gen_attr_2`, `col_1` AS `gen_attr_0`, `ts` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_0` GROUP BY `gen_attr_0`")
+        else
+            sql8 should be ("SELECT `gen_attr_0` AS `col_1`, sum(`gen_attr_2`) AS `sum` FROM (SELECT `col_0` AS `gen_attr_2`, `col_1` AS `gen_attr_0`, `ts` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_0` GROUP BY `gen_attr_0`")
+
+    })
+
+    it should "support GROUP BY expressions with cardinal expressions" in (if (hiveSupported) {
+        val df8 = spark.sql(
+            """
+              |SELECT
+              |   col_1,
+              |   sum(col_0) AS sum
+              |FROM sql_builder_0
+              |GROUP BY 1
+              |""".stripMargin)
+        val sql8 = new SqlBuilder(df8.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql8))
+        if (spark.version < "3.2")
+            sql8 should be ("SELECT `gen_attr_0` AS `col_1`, sum(CAST(`gen_attr_2` AS BIGINT)) AS `sum` FROM (SELECT `col_0` AS `gen_attr_2`, `col_1` AS `gen_attr_0`, `ts` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_0` GROUP BY `gen_attr_0`")
+        else
+            sql8 should be ("SELECT `gen_attr_0` AS `col_1`, sum(`gen_attr_2`) AS `sum` FROM (SELECT `col_0` AS `gen_attr_2`, `col_1` AS `gen_attr_0`, `ts` AS `gen_attr_3` FROM `default`.`sql_builder_0`) AS `gen_subquery_0` GROUP BY `gen_attr_0`")
     })
 
     it should "support sort ordering" in (if (hiveSupported) {
@@ -365,7 +419,20 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |ORDER BY col_0""".stripMargin)
         val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql1))
-        sql1.withoutQuotes should be ("SELECT gen_attr_0 AS col_0, gen_attr_1 AS col_1 FROM (SELECT col_0 AS gen_attr_0, col_1 AS gen_attr_1 FROM default.sql_builder_0) AS gen_subquery_1 ORDER BY gen_attr_0 ASC")
+        sql1 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1` FROM `default`.`sql_builder_0`) AS `gen_subquery_1` ORDER BY `gen_attr_0` ASC")
+    })
+
+    it should "support sort ordering with cardinal expressions" in (if (hiveSupported) {
+        val df1 = spark.sql(
+            """
+              |SELECT
+              |  x.col_0,
+              |  x.col_1
+              |FROM sql_builder_0 x
+              |ORDER BY 1""".stripMargin)
+        val sql1 = new SqlBuilder(df1.queryExecution.analyzed).toSQL
+        noException shouldBe thrownBy(spark.sql(sql1))
+        sql1 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1` FROM `default`.`sql_builder_0`) AS `gen_subquery_1` ORDER BY `gen_attr_0` ASC")
     })
 
     it should "support sort ordering with DESC NULLS FIRST" in (if (hiveSupported) {
@@ -378,7 +445,7 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |ORDER BY col_0 DESC NULLS FIRST""".stripMargin)
         val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql2))
-        sql2.withoutQuotes should be ("SELECT gen_attr_0 AS col_0, gen_attr_1 AS col_1 FROM (SELECT col_0 AS gen_attr_0, col_1 AS gen_attr_1 FROM default.sql_builder_0) AS gen_subquery_1 ORDER BY gen_attr_0 DESC NULLS FIRST")
+        sql2 should be ("SELECT `gen_attr_0` AS `col_0`, `gen_attr_1` AS `col_1` FROM (SELECT `col_0` AS `gen_attr_0`, `col_1` AS `gen_attr_1` FROM `default`.`sql_builder_0`) AS `gen_subquery_1` ORDER BY `gen_attr_0` DESC NULLS FIRST")
     })
 
     it should "support VARCHAR(n) and CHAR(n) columns" in (if (hiveSupported && hiveVarcharSupported) {
@@ -390,6 +457,6 @@ class SqlBuilderTest extends AnyFlatSpec with Matchers with LocalSparkSession {
               |FROM sql_builder_2 x""".stripMargin)
         val sql2 = new SqlBuilder(df2.queryExecution.analyzed).toSQL
         noException shouldBe thrownBy(spark.sql(sql2))
-        sql2.withoutQuotes should be("SELECT col_0, col_1 FROM default.sql_builder_2")
+        sql2 should be("SELECT `col_0`, `col_1` FROM `default`.`sql_builder_2`")
     })
 }