+++ docs/releases.md
@@ -30,6 +30,11 @@ Breaking changes will be documented in the changelog file for each version.
 The following gives an (incomplete) list of past releases of the last 12 months. This will help you to spot breaking 
 changes over time.
 
+### Version 1.3.1 - 2024-12-16
+
+* github-514: Generated SQL code for views should quote all column names
+
+
 ### Version 1.3.0 - 2024-12-06
 
 * github-510: Upgrade Cloudera Spark 3.3 build profile